fix: 코드 리뷰에서 발견된 버그 수정 및 로컬 Flyway 전환
- /lookup 잘못된 요청 파라미터(누락/타입 오류) 시 500 대신 400 반환하도록 예외 핸들러 추가
- 행정동 경계선 위 좌표 매칭 실패 문제 수정 (contains -> covers)
- 폴리곤 홀(구멍) 무시하던 GeoJSON 파싱 로직 수정 및 중복 제거
- 로컬 ddl-auto=create가 docker 목업 데이터를 매 기동마다 지우던 문제 해결:
  Flyway를 로컬에서도 사용하도록 전환, db/init 중복 스키마/목업 SQL을 V17 마이그레이션으로 통합
- 다운로드 실패로 손상된 미사용 hangjeongdong.geojson 삭제
- store_lat/store_lng 컬럼 타입을 엔티티(Double)와 일치하도록 DOUBLE로 수정 (validate 모드에서 스키마 불일치 발견)
- JUnit5 + AssertJ 단위 테스트 추가 (GlobalExceptionHandlerTest, RegionLookupServiceTest)

docker compose up + 실제 앱 기동으로 마이그레이션/엔드포인트 동작 검증 완료

+++ src/main/resources/db/migration/V16__add_store_coordinates_to_users.sql
@@ -1,3 +1,3 @@
 ALTER TABLE users
-    ADD COLUMN store_lat DECIMAL(10, 7) NULL COMMENT '가게 위도' AFTER store_region,
-    ADD COLUMN store_lng DECIMAL(10, 7) NULL COMMENT '가게 경도' AFTER store_lat;
+    ADD COLUMN store_lat DOUBLE NULL COMMENT '가게 위도' AFTER store_region,
+    ADD COLUMN store_lng DOUBLE NULL COMMENT '가게 경도' AFTER store_lat;
